forum v3.2: like and unlike forum comment, divide list forum into private and public

+++ src/main/java/org/safescan/service/impl/ForumServiceImpl.java
@@ -2,6 +2,7 @@ package org.safescan.service.impl;
 
 import org.safescan.DTO.ForumDTO;
 import org.safescan.DTO.UserForumDTO;
+import org.safescan.mapper.ForumCommentMapper;
 import org.safescan.mapper.ForumMapper;
 import org.safescan.service.ForumService;
 import org.safescan.utils.ThreadLocalUtil;
@@ -11,12 +12,16 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @Service
 public class ForumServiceImpl implements ForumService {
 
     @Autowired
     private ForumMapper forumMapper;
+
+    @Autowired
+    private ForumCommentMapper forumCommentMapper;
 
     @Override
     public void add(ForumDTO forum) {
@@ -37,9 +42,16 @@ public class ForumServiceImpl implements ForumService {
     }
 
     @Override
-    public List<UserForumDTO> getForums(int page, int size) {
+    public List<UserForumDTO> getForums(int page, int size, Integer userId) {
         int offset = page * size; // Calculating the offset
-        return forumMapper.getForums(offset, size);
+        List<UserForumDTO> userForums = forumMapper.getForums(offset, size);
+
+        for (UserForumDTO userForum : userForums) {
+            // If userId is null, it is public list method
+            boolean isLiked = userId != null && forumMapper.isLikedByUserId(userId, userForum.getForumId());
+            userForum.setLiked(isLiked);
+        }
+        return userForums;
     }
 
     @Override
@@ -51,26 +63,35 @@ public class ForumServiceImpl implements ForumService {
     @Override
     public void delete(Integer forumId) {
         LocalDateTime updateTime = LocalDateTime.now();
-        forumMapper.delete(forumId, updateTime);
-    }
-
-    @Override
-    public void likeForum(Integer forumId, Integer userId) {
-        LocalDateTime time = LocalDateTime.now();
-        forumMapper.addForumLikes(forumId, userId, time);
-        forumMapper.updateForumLikes(forumId, time, "add");
-    }
-
-    @Override
-    public void unlikeForum(Integer forumId, Integer userId) {
-        LocalDateTime time = LocalDateTime.now();
-        forumMapper.deleteForumLikes(forumId, userId);
-        forumMapper.updateForumLikes(forumId, time, "delete");
+        forumMapper.deleteByForumId(forumId, updateTime);
+        forumCommentMapper.deleteByForumCommentId(forumId,null, updateTime);
     }
 
     @Override
     public ForumDTO getLikedForum(Integer forumId, Integer userId) {
         return forumMapper.getLikedForum(forumId, userId);
+    }
+
+    @Override
+    public void likeForum(Integer forumId, Integer userId) {
+        LocalDateTime time = LocalDateTime.now();
+
+        if (Objects.equals(forumMapper.getStateByForumId(forumId), "Deleted")) {
+            throw new RuntimeException("This post has been deleted!");
+        }
+        forumMapper.addForumLikes(forumId, userId, time);
+        forumMapper.updateForumLikes(forumId, time, "like");
+    }
+
+    @Override
+    public void unlikeForum(Integer forumId, Integer userId) {
+        LocalDateTime time = LocalDateTime.now();
+
+        if (Objects.equals(forumMapper.getStateByForumId(forumId), "Deleted")) {
+            throw new RuntimeException("This post has been deleted!");
+        }
+        forumMapper.deleteForumLikes(forumId, userId);
+        forumMapper.updateForumLikes(forumId, time, "unlike");
     }
